+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/preparator/extension/CertificateTypeExtensionPreparatorTest.java
@@ -44,9 +44,9 @@ public class CertificateTypeExtensionPreparatorTest {
         preparator.prepare();
 
         assertArrayEquals(ExtensionType.CERT_TYPE.getValue(), msg.getExtensionType().getValue());
-        assertEquals(extensionLength, (int) msg.getExtensionLength().getValue());
+        assertEquals(extensionLength, (long) msg.getExtensionLength().getValue());
         assertArrayEquals(CertificateType.toByteArray(certList), msg.getCertificateTypes().getValue());
-        assertEquals(cerListLength, (int) msg.getCertificateTypesLength().getValue());
+        assertEquals(cerListLength, (long) msg.getCertificateTypesLength().getValue());
     }
 
 }
