+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/preparator/extension/ClientAuthzExtensionPreparatorTest.java
@@ -41,7 +41,7 @@ public class ClientAuthzExtensionPreparatorTest {
         preparator.prepare();
 
         assertArrayEquals(ExtensionType.CLIENT_AUTHZ.getValue(), msg.getExtensionType().getValue());
-        assertEquals(authzFormatListLength, (int) msg.getAuthzFormatListLength().getValue());
+        assertEquals(authzFormatListLength, (long) msg.getAuthzFormatListLength().getValue());
         assertArrayEquals(authzFormatListAsBytes, msg.getAuthzFormatList().getValue());
     }
 }
